% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmsMB-class.R
\name{tmsMB}
\alias{tmsMB}
\title{tmsMB}
\description{
build and break gene regulatory modesl
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{identifier for a class object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{identifier for a class object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-tmsMB-new}{\code{tmsMB$new()}}
\item \href{#method-tmsMB-setStudyRegion}{\code{tmsMB$setStudyRegion()}}
\item \href{#method-tmsMB-getStudyRegion}{\code{tmsMB$getStudyRegion()}}
\item \href{#method-tmsMB-getFimoGenomicRegion}{\code{tmsMB$getFimoGenomicRegion()}}
\item \href{#method-tmsMB-get.gtex.eqtls}{\code{tmsMB$get.gtex.eqtls()}}
\item \href{#method-tmsMB-get.ampad.eqtls}{\code{tmsMB$get.ampad.eqtls()}}
\item \href{#method-tmsMB-getGTEx.eqtl.genomicRegion}{\code{tmsMB$getGTEx.eqtl.genomicRegion()}}
\item \href{#method-tmsMB-getAMPAD.eqtl.genomicRegion}{\code{tmsMB$getAMPAD.eqtl.genomicRegion()}}
\item \href{#method-tmsMB-getGTEx.eqtl.tissues}{\code{tmsMB$getGTEx.eqtl.tissues()}}
\item \href{#method-tmsMB-get.current.GTEx.eqtl.tissue}{\code{tmsMB$get.current.GTEx.eqtl.tissue()}}
\item \href{#method-tmsMB-set.current.GTEx.eqtl.tissue}{\code{tmsMB$set.current.GTEx.eqtl.tissue()}}
\item \href{#method-tmsMB-run.tms}{\code{tmsMB$run.tms()}}
\item \href{#method-tmsMB-add.eqtls.toTmsTable}{\code{tmsMB$add.eqtls.toTmsTable()}}
\item \href{#method-tmsMB-get.tmsTable}{\code{tmsMB$get.tmsTable()}}
\item \href{#method-tmsMB-get.knownSnps}{\code{tmsMB$get.knownSnps()}}
\item \href{#method-tmsMB-get.motifBreaks}{\code{tmsMB$get.motifBreaks()}}
\item \href{#method-tmsMB-get.breaksTable}{\code{tmsMB$get.breaksTable()}}
\item \href{#method-tmsMB-run.trena}{\code{tmsMB$run.trena()}}
\item \href{#method-tmsMB-get.trenaTable}{\code{tmsMB$get.trenaTable()}}
\item \href{#method-tmsMB-set.tmsFilteredTable}{\code{tmsMB$set.tmsFilteredTable()}}
\item \href{#method-tmsMB-get.tmsFilteredTable}{\code{tmsMB$get.tmsFilteredTable()}}
\item \href{#method-tmsMB-breakMotifs}{\code{tmsMB$breakMotifs()}}
\item \href{#method-tmsMB-viz}{\code{tmsMB$viz()}}
\item \href{#method-tmsMB-clone}{\code{tmsMB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-new"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$new(
  targetGene,
  trenaProject,
  tbl.fimo,
  tbl.gtex.eqtls,
  tbl.ampad.eqtls,
  tbl.oc,
  known.snps,
  chromosome = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetGene}}{the gene whose regulation we wish to explore}

\item{\code{trenaProject}}{a TrenaProject subclass, for access to, e.g., expression data}

\item{\code{tbl.fimo}}{results from a broad low-filter fimo run}

\item{\code{tbl.gtex.eqtls}}{eQTLs for a specific tissue}

\item{\code{tbl.ampad.eqtls}}{eQTLs for ROSMAP and Mayo}

\item{\code{tbl.oc}}{open chromatin}

\item{\code{known.snps}}{genomic variants already looked up}

\item{\code{chromosome}}{chromosome id, usually not needed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of the class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-setStudyRegion"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-setStudyRegion}{}}}
\subsection{Method \code{setStudyRegion()}}{
establish the genomic region of interest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$setStudyRegion(chrom, start, end)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chrom}}{chromosome}

\item{\code{start}}{region start}

\item{\code{end}}{region end}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
noting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-getStudyRegion"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-getStudyRegion}{}}}
\subsection{Method \code{getStudyRegion()}}{
obtain the genomic region of interest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$getStudyRegion()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-getFimoGenomicRegion"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-getFimoGenomicRegion}{}}}
\subsection{Method \code{getFimoGenomicRegion()}}{
obtain the full extent of the fimo table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$getFimoGenomicRegion()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-get.gtex.eqtls"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-get.gtex.eqtls}{}}}
\subsection{Method \code{get.gtex.eqtls()}}{
obtain the eqtls for current GTEx tissue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$get.gtex.eqtls()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-get.ampad.eqtls"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-get.ampad.eqtls}{}}}
\subsection{Method \code{get.ampad.eqtls()}}{
obtain the eqtls ROSMAP and Mayo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$get.ampad.eqtls()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-getGTEx.eqtl.genomicRegion"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-getGTEx.eqtl.genomicRegion}{}}}
\subsection{Method \code{getGTEx.eqtl.genomicRegion()}}{
obtain the genomic extent of the GTEx eqtls
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$getGTEx.eqtl.genomicRegion()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-getAMPAD.eqtl.genomicRegion"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-getAMPAD.eqtl.genomicRegion}{}}}
\subsection{Method \code{getAMPAD.eqtl.genomicRegion()}}{
obtain the genoic extent of the ampad eqtls
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$getAMPAD.eqtl.genomicRegion()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-getGTEx.eqtl.tissues"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-getGTEx.eqtl.tissues}{}}}
\subsection{Method \code{getGTEx.eqtl.tissues()}}{
obtain the full list of GTEx tissue names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$getGTEx.eqtl.tissues()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-get.current.GTEx.eqtl.tissue"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-get.current.GTEx.eqtl.tissue}{}}}
\subsection{Method \code{get.current.GTEx.eqtl.tissue()}}{
obtain the current GTEx tissule name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$get.current.GTEx.eqtl.tissue()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-set.current.GTEx.eqtl.tissue"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-set.current.GTEx.eqtl.tissue}{}}}
\subsection{Method \code{set.current.GTEx.eqtl.tissue()}}{
specify the current GTEx tissue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$set.current.GTEx.eqtl.tissue(new.tissue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new.tissue}}{a recognized GTEx tissue name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-run.tms"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-run.tms}{}}}
\subsection{Method \code{run.tms()}}{
run the full TrenaMultiScore process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$run.tms()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-add.eqtls.toTmsTable"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-add.eqtls.toTmsTable}{}}}
\subsection{Method \code{add.eqtls.toTmsTable()}}{
add eqtls values to the tms feature table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$add.eqtls.toTmsTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-get.tmsTable"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-get.tmsTable}{}}}
\subsection{Method \code{get.tmsTable()}}{
obtain the current feature table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$get.tmsTable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-get.knownSnps"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-get.knownSnps}{}}}
\subsection{Method \code{get.knownSnps()}}{
obtain our current list of location-specified variants
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$get.knownSnps()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-get.motifBreaks"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-get.motifBreaks}{}}}
\subsection{Method \code{get.motifBreaks()}}{
obtain the results from motifbreakR
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$get.motifBreaks()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-get.breaksTable"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-get.breaksTable}{}}}
\subsection{Method \code{get.breaksTable()}}{
obtain the data.frame summary out of motifbreakR
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$get.breaksTable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-run.trena"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-run.trena}{}}}
\subsection{Method \code{run.trena()}}{
build trena regulatory model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$run.trena(tf.candidates)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tf.candidates}}{character vector,  TFs selected from the feature table}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-get.trenaTable"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-get.trenaTable}{}}}
\subsection{Method \code{get.trenaTable()}}{
obtain the trena results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$get.trenaTable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-set.tmsFilteredTable"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-set.tmsFilteredTable}{}}}
\subsection{Method \code{set.tmsFilteredTable()}}{
user (biologist) chosen subset of the feature table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$set.tmsFilteredTable(tbl.filtered)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl.filtered}}{user (biologist) filtered feature table}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-get.tmsFilteredTable"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-get.tmsFilteredTable}{}}}
\subsection{Method \code{get.tmsFilteredTable()}}{
obtain the chosen filtered subset of the feature table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$get.tmsFilteredTable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-breakMotifs"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-breakMotifs}{}}}
\subsection{Method \code{breakMotifs()}}{
run motifbreakR
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$breakMotifs(tbl.trena, tbl.tms, tbl.eqtls)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl.trena}}{trena results}

\item{\code{tbl.tms}}{feature table}

\item{\code{tbl.eqtls}}{eQTLs of interest}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-viz"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-viz}{}}}
\subsection{Method \code{viz()}}{
render breaks, eQTLs and trena model in an existing igvR instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$viz(
  igv,
  current.tissue,
  tbl.trena,
  tbl.tms,
  tbl.tms.filtered,
  tbl.gtex.eqtls,
  tbl.ampad.eqtls,
  tbl.oc,
  tbl.breaks
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{igv}}{igvR instance}

\item{\code{current.tissue}}{GTEx tissue of interest}

\item{\code{tbl.trena}}{results of trena run}

\item{\code{tbl.tms}}{the full feature table}

\item{\code{tbl.tms.filtered}}{the user-filtered feature table}

\item{\code{tbl.gtex.eqtls}}{GTEx eqtls}

\item{\code{tbl.ampad.eqtls}}{ROSMAP and Mayo eqtls}

\item{\code{tbl.oc}}{open chromatin}

\item{\code{tbl.breaks}}{motifbreakR results}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmsMB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-tmsMB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmsMB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
